+++ app/routes/_app.blog_.$slug/article-utils.ts
@@ -1,0 +1,9 @@
+// Utility function to generate heading IDs
+export function generateHeadingId(text: string): string {
+  return text
+    .toLowerCase()
+    .replace(/[^a-z0-9\s-]/g, "")
+    .replace(/\s+/g, "-")
+    .replace(/-+/g, "-")
+    .replace(/^-|-$/g, "");
+}
